+++ CHANGELOG.md
@@ -4,7 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 # [8.9.2] - 2024-07-12
 - Refactoring to accommodate using v2.0.0 of Vonage JWT library
-- 
+- Fixed `com.vonage.client.voice.EventWebhook` deserialisation issue
+  - `getCallUuid()` and `getRecordingUuid()` now return String instead of UUID
 
 # [8.9.1] - 2024-07-09
 - Fixed parsing issue in `ConversationsClient#listEvents`